chore: move helper methods to Test Message Service
